Refactor problem form and add utility functions
Refactored the problem form component, extracted logic into a composable, and added utility functions for better code organization and reusability.
