+++ src/components/AdminDashboard/AdminProblems/ProblemForm/ProblemContentEditor.vue
